# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "service for distributing test files among processes/machines/containers",
        "title": "Split specs API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "ovr.shevtsov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "parallel-specs.shelex.dev",
    "basePath": "/",
    "paths": {
...
        "/api/keys": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "api key"
                ],
                "summary": "get user api keys",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "api keys",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.ApiKey"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "api key"
                ],
                "summary": "add new api key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ApiKeyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "api token",
                        "schema": {
                            "$ref": "#/definitions/controllers.tokenResponse"
                        }
                    }
                }
            }
        },
        "/api/keys/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "api key"
                ],
                "summary": "delete api key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "api key id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/projects": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "get projects for user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "projects",
                        "schema": {
                            "$ref": "#/definitions/controllers.ProjectsResponse"
                        }
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "delete project by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/projects/{id}/sessions": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "get project recorded sessions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pagination",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sessions",
                        "schema": {
                            "$ref": "#/definitions/controllers.ProjectSessions"
                        }
                    }
                }
            }
        },
        "/api/projects/{id}/share/{email}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "share project with another user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "other account email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.ApiKeyInput": {
            "type": "object",
            "required": [
                "expireAt",
                "name"
            ],
            "properties": {
                "expireAt": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "controllers.ProjectSessions": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Session"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "controllers.ProjectsResponse": {
            "type": "object",
            "properties": {
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Project"
                    }
                }
            }
        },
...
        "controllers.tokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "entities.ApiKey": {
            "type": "object",
            "properties": {
                "expireAt": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "entities.Execution": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "estimatedDuration": {
                    "type": "integer"
                },
                "finishedAt": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "machineId": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                },
                "specId": {
                    "type": "string"
                },
                "specName": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "entities.Project": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entities.Session": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "integer"
                },
                "executions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Execution"
                    }
                },
                "finishedAt": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "integer"
                }
            }
        },
...
    }
}